requirements.txt
asgiref==3.5.1
Django==4.0.4
python-slugify==6.1.2
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
